backend/pyproject.toml
[project]
name = "sonic-backend"
version = "1.0.0"
description = "Sonic Backend API"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "psycopg2-binary>=2.9.9",
    "pillow>=10.0.0",
    "python-decouple>=3.8",
    "django-filter>=23.3",
    "drf-spectacular>=0.27.0",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "daphne>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sonic_backend", "sonic_app"]


